Convert python to use native libmagic python-magic bindings
There are two Python modules calling themselves 'python-magic'.
One is part of the upstream file / libmagic project (this one is
shipped in Fedora and other Red Hat distros). The other is a
standalone cPython wrapper that has a different API.

Convert the SOSCleaner code to use the libmagic bindings when
determining file types.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>