Adding some default ubuntu users in the ignored_user list
These users will be visible inside the soscleaner tarball, but they
don't reveal anything important IMHO. They are already known users
deployed by default during a ubuntu container launch for instance.
I added them to the ignore list because they can be easily
misinterpreted.

Example: 'daemon', 'sys', 'syslog' 'sshd' are username, but can also be
a process, systemd service, a keyword in the logs, a filename, and so on.

To avoid confusion, I think it's preferable to ignore them, until we
come up with a better solution such as refinning the regex or else.

Signed-off-by: Eric Desrochers <eric.desrochers@canonical.com>