Adding a good debian-ubuntu-cloud user base in ignored list
Fix: #149

We are limited in what we can do with user obfuscation due to
the fact that soscleaner can only rely on lastlog. A good thing
would have been to be able to filter by uid > 1000 to avoid
obfuscating well known system users as the focus here should
be on users created by users (internal users). Adding this
significant ignored_users list will be beneficial for soscleaner
to spend time on obfuscating unecessary users, avoid some false
positives situation and let soscleaner focus more on users created by
users. Unfortunately there will always have corner cases, but after
turning the question upsidedown and discuss with peers, it seems to
be the way to go at least for now, until soscleaner is refine or
sosreport to provide a better way to collect users.

Signed-off-by: Eric Desrochers <eric.desrochers@canonical.com>